+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+# [0.9.1] (2025-03-05)
+
 ## Changes
 
 -   Add migration.js script to automate database updates with Ansible